refactor ocsp test
