+++ src/test/groovy/kz/ncanode/common/WithTestData.groovy
@@ -2,6 +2,11 @@ package kz.ncanode.common
 
 import kz.ncanode.constants.MessageConstants
 import kz.ncanode.dto.request.SignerRequest
+import org.apache.http.HttpEntity
+import org.apache.http.client.methods.CloseableHttpResponse
+
+import static org.mockito.Mockito.doReturn
+import static org.mockito.Mockito.mock
 
 interface WithTestData {
 
@@ -80,4 +85,14 @@ interface WithTestData {
     final static String NCA_2004_RSA_CERT = 'LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tDQpNSUlHYWpDQ0JGS2dBd0lCQWdJVVJTcGtOMkVjR0QrL0hzV2h1a2tnM1hJVWlwc3dEUVlKS29aSWh2Y05BUUVMDQpCUUF3TmpFbk1DVUdBMVVFQXd3ZTBKM1FsZENUMEliUWw5Q1QwSVlnMEpyUW5pQlVSVk5VSUNoU1UwRXBNUXN3DQpDUVlEVlFRR0V3SkxXakFlRncweE9UQXpNVFV3TXpReU1UZGFGdzB5TkRBek1UTXdNelF5TVRkYU1DMHhDekFKDQpCZ05WQkFZVEFrdGFNUjR3SEFZRFZRUUREQlhTc05DYTBKNGdNeTR3SUNoU1UwRWdWRVZUVkNrd2dnSWlNQTBHDQpDU3FHU0liM0RRRUJBUVVBQTRJQ0R3QXdnZ0lLQW9JQ0FRRFUyL1ZFSGZhQThJbXoyY0puK3FpcTR2Q3FOTmNLDQpnQ280Mm1MVDJhbHJJajhlajM0TytZNytHMWR5VGVmZlZZUWpHakRnTHB0V292Z1RSMVBkbWN5SnNGeDJkTGJJDQp0bk9PRUF6OUFTNnpUM1JVY3RnaXU5T0dDbFM2SXptOUx1RTUxMzRKYmpVZjFlVXFCNVo0ekx6RTBncXJ1MGV6DQpvWmoxdVNDVTRvRnZRd0dKcjhqZDZIczNxR3N4cGRHQWxMSTN3WHVqTkdWK1JFWGdNRGdGSW50MUhrdS81UHBTDQpndlJFMFVYYVF3NHQ1MEsrU2gxN1loWkJ4YnMyMk5nNER2d2VrYS9BYXI3QVFFSGZmOCtUcWgzdzZUUDRhV1BDDQpJUnRpcHhqYVVjNWtHVjYzRXc5OVQ5RmxscEt0ZVh2d1VRSEUxclRDSThxL01QcS82MElBbWJ2SnBDRk4zamtMDQpVakZWRE15SCtCbFhRNFJiMFRqU1VPdndHMlpoMTZJOWhPVmlZQWdBbVozWWdjNElyTy9XSFlWUVpjLzQ3bDNBDQpHSWR4UWlaUFppdDhMWWVLUnViZ1NOdGxUUGd3RXRvaUpQQkE1eENPQzhkNGhoZWtJaUtMNlZYVWVnUk9ra0QyDQpzbXlTalREZGNybks5Y0lkQi8vZ2RPOVJMZ053b2xpK09IQ3VSdXdNQVdudXdOUVNLWnVReVhyT25ETy9RcndsDQptalpIa2JtT2IwcUlpdHBGRFg3aTh0cmdyNy9hTXlxUUQvTVUvcmpLNEtNVzR0em9XUzJmanZhSk5ZSUZGZERxDQpDSzBIUEd0aWh5RDU1R1M5WGplV2FieTBlRHFNOU9VSHZud1J6N1RpUDlYejJSSGxBZXRkV081YTN4c0dzckNDDQp2OTE0RmFub3A5Ymtsd0lEQVFBQm80SUJkekNDQVhNd0hRWURWUjBPQkJZRUZLYU1Gak44dU9nMVp3WStYa0ZYDQpWYUt2TkZCb01IRUdBMVVkSXdScU1HaUFGTlBIZmYrc292VU93V01aVzYvK2NuZFhzanppb1Rxa09EQTJNU2N3DQpKUVlEVlFRRERCN1FuZENWMEpQUWh0Q1gwSlBRaGlEUW10Q2VJRlJGVTFRZ0tGSlRRU2t4Q3pBSkJnTlZCQVlUDQpBa3RhZ2hSVHgzMy9yS0wxRHNGakdWdXYvbkozVjdJODRqQVBCZ05WSFJNQkFmOEVCVEFEQVFIL01BNEdBMVVkDQpEd0VCL3dRRUF3SUJCakJKQmdnckJnRUZCUWNCQVFROU1Ec3dPUVlJS3dZQkJRVUhNQUtHTFdoMGRIQTZMeTkwDQpaWE4wTG5CcmFTNW5iM1l1YTNvdlkyVnlkQzl5YjI5MFgzSnpZVjkwWlhOMExtTmxjakE1QmdOVkhTQUVNakF3DQpNQzRHQnlxRERnTURBUUV3SXpBaEJnZ3JCZ0VGQlFjQ0FSWVZhSFIwY0RvdkwzQnJhUzVuYjNZdWEzb3ZZM0J6DQpNRGdHQTFVZEh3UXhNQzh3TGFBcm9DbUdKMmgwZEhBNkx5OTBaWE4wTG5CcmFTNW5iM1l1YTNvdlkzSnNMM0p2DQpiM1JmY25OaExtTnliREFOQmdrcWhraUc5dzBCQVFzRkFBT0NBZ0VBcU9zRXIrYmUrcHJOL2NJbm1NR1BmTDhRDQpIaEN3NndHMG5QRUd4UTg5VitkejlxbXBnaEZnQzNEdTFGUm0reDFadEo5dFhRRHA5MFdaWXBGeXc1bkVEeW9RDQpXSUt5eWFqNmwxUHVRbnZqOHhLbGErRmJSb3ZzRFYrZzg1SkEwNWI1bGtvS1JGdFVCY0g5L1ZXM2NKNFBMUUtlDQpqR2F3NzNoZldycEtuU2FJL2JyYnRNMnVpaWM2VTA4N09TcUtpTVRlMW93TTVXUXhMZzBtVmdoMVBaL3dOMnZqDQpIU0x1N0dYajYyK3VnUHhocjhGaFRmQ0Y4MHZldkxMbW9Jc1JqVTYzNUp5blNKaHU0VUlhZ1dCSFA0MkJsTHZODQppcXpEVTVzeG8vTjByd0QxNzEvajIxWGNxYU9rdHBGdWpqWS8rWkgyTUx5VVZLWkIrUDMyTGhNck5EcTlaY2RoDQppdTBXcXZ3dXZHZHN6dlQ3U252dERQVlVVSzk2ZmRDWjQ0K0NZN0VaeVhMY0Z5UFhsTVI2VUZuSXJXSnZTMVNZDQpvSnNPeEdtMTV2b2ZEanFZbXY4aldnRVpGSWZmcDc0ZEx4V0s2UTJGUHdxWHNtVDN0eVYrTkROTnRKbzV3bmIvDQpFMjR0ZDBSL1loclhiVDFjV3BRQlFzcFFpY1BWczloNjBvYXkzakx2aUMrak5IbTV1M0lwbEsvc040QWtUSXJyDQpkNmNPdlVGVm1STXorK0pGQ1BWUGlLOXNPUExITzNmUEk1UjBCQ3NqNHlRbTlDdlhCK3J2ME9nM2xFbE9oWE5UDQpiRUN6WU5BZHc2OG1WZTNyNVJkcWJKOTJnS2h1UGdlRTY4RmpTS0VBUktMaWRDMnhMVUJEYXpIMlgvdFBZSU0rDQpGQTE3eG1Rd0dXWHNmZndwSzEwPQ0KLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQ0K'
     final static String NCA_2015_CERT = 'MIID/TCCA2WgAwIBAgIUetJLG6OgyWH+HKhQPmqiu0UNuKMwDgYKKoMOAwoBAQIDAgUAMF8xUDBOBgNVBAMMR9Cd0JXQk9CG0JfQk9CGINCa0KPTmNCb0JDQndCU0KvQoNCj0KjQqyDQntCg0KLQkNCb0KvSmiAoR09TVCkgVEVTVCAyMDIyMQswCQYDVQQGEwJLWjAeFw0yMjA3MDcwOTU1MjNaFw0zMjA3MDQwOTU1MjNaMF0xTjBMBgNVBAMMRdKw0JvQotCi0KvSmiDQmtCj05jQm9CQ0J3QlNCr0KDQo9Co0Ksg0J7QoNCi0JDQm9Cr0pogKEdPU1QpIFRFU1QgMjAyMjELMAkGA1UEBhMCS1owgawwIwYJKoMOAwoBAQICMBYGCiqDDgMKAQECAgEGCCqDDgMKAQMDA4GEAASBgLg6ffxI/sxPLaDh3hLsK8f9P7PMFjc56V21F+SDFKNkURgSXG93YnmPUuKv6NqP8flHGTFV+gpGiM4p3XPRbdE3sg0hQ8SaoHXR2h6hofM3OycgAFJHIBHDsrM5McqfgkuVVgnP/FnUkzmEG3/sC0Z9c7Ux27UsnudfeVvVEBhro4IBpzCCAaMwDwYDVR0TAQH/BAUwAwEB/zAOBgNVHQ8BAf8EBAMCAQYwOgYDVR0fBDMwMTAvoC2gK4YpaHR0cDovL2NybC5yb290Lmdvdi5rei9nb3N0X3Rlc3RfMjAyMi5jcmwwHQYDVR0OBBYEFPrSSxujoMlh/hyoUD5qortFDbijMIGcBgNVHSMEgZQwgZGAFKfTq3eQNy5baYFw3eYMzKOLw6SyoWOkYTBfMVAwTgYDVQQDDEfQndCV0JPQhtCX0JPQhiDQmtCj05jQm9CQ0J3QlNCr0KDQo9Co0Ksg0J7QoNCi0JDQm9Cr0pogKEdPU1QpIFRFU1QgMjAyMjELMAkGA1UEBhMCS1qCFCfTq3eQNy5baYFw3eYMzKOLw6SyMEsGCCsGAQUFBwEBBD8wPTA7BggrBgEFBQcwAoYvaHR0cDovL3Jvb3QuZ292Lmt6L2NlcnQvcm9vdF90ZXN0X2dvc3RfMjAyMi5jZXIwOQYDVR0gBDIwMDAuBgYqgw4DAwIwJDAiBggrBgEFBQcCARYWIGh0dHA6Ly9wa2kuZ292Lmt6L2NwczAOBgoqgw4DCgEBAgMCBQADgYEAF1XLtw6Fcdzrl33J2DaP125p+nQviyEddDJxb0mUlxbWUTAZlKsAqphfN/SxEd2w0pJL39hTysG0rAC6Lp5bdMdmsclPZqdn22SqiCluNXUpTj84jh8hdT9QVB6Ll4yYlBf9WlZUFqzWF6oR151h1xL1D1wcq8c2SQpMUZFOKHU='
 
+
+    default CloseableHttpResponse createMockedResponse(InputStream inputStream) {
+        def entity = mock(HttpEntity)
+        doReturn(inputStream).when(entity).getContent()
+
+        def response = mock(CloseableHttpResponse)
+        doReturn(entity).when(response).getEntity()
+
+        return response
+    }
 }
